+++ locan/scripts/script_draw_roi.py
@@ -25,7 +25,7 @@ import argparse
 from pathlib import Path
 
 from locan.gui.io import file_dialog
-import locan.io.io_locdata as io
+import locan.io.locdata.io_locdata as io
 from locan.render.render2d import select_by_drawing_mpl