+++ locan/scripts/script_napari.py
@@ -27,7 +27,7 @@ if _has_napari: import napari
 
 from locan.constants import FileType
 from locan.gui.io import file_dialog
-import locan.io.io_locdata as io
+import locan.io.locdata.io_locdata as io
 from locan.render.render2d import render_2d_napari