+++ locan/tests/data/test_transform.py
@@ -2,12 +2,10 @@ import pytest
 import numpy as np
 import matplotlib.pyplot as plt
 
-from locan import LocData
 import locan.constants
 from locan.constants import _has_open3d
 from locan.data.region import Polygon
-from locan import render_2d
-from locan.io.io_locdata import load_rapidSTORM_file
+from locan.io.locdata.io_locdata import load_rapidSTORM_file
 from locan.data.transform import randomize, transform_affine
 from locan.data.transform.transformation import _homogeneous_matrix
 from locan.data.transform.bunwarpj import _read_matrix, _unwarp, bunwarp